appBureau/requirements.txt
customtkinter
packaging
pillow
bcrypt
sqlalchemy # DBlayer
pymysql # Driver MySQL pour SQLAlchemy
python-dotenv # Lecture .env
uuid